add self implemented json utils
